@@ -39,46 +39,46 @@
   },
   "homepage": "https://github.com/wikimedia/mathoid",
   "dependencies": {
-    "bluebird": "^3.5.1",
-    "body-parser": "^1.18.2",
+    "bluebird": "^3.7.2",
+    "body-parser": "^1.19.0",
     "bunyan": "^1.8.12",
     "cassandra-uuid": "^0.1.0",
-    "compression": "^1.7.1",
-    "express": "^4.16.2",
-    "http-shutdown": "^1.2.0",
-    "js-yaml": "^3.12.2",
+    "compression": "^1.7.4",
+    "express": "^4.17.1",
+    "http-shutdown": "^1.2.2",
+    "js-yaml": "^3.13.1",
     "mathoid-mathjax-node": "^0.7.0",
     "mock-express-request": "^0.2.2",
     "mock-express-response": "^0.2.2",
-    "preq": "^0.5.4",
-    "service-runner": "^2.6.17",
-    "speech-rule-engine": "^2.3.2",
-    "svgo": "^1.2.0",
-    "swagger-router": "^0.7.1",
-    "texvcinfo": "^0.5.4"
+    "preq": "^0.5.14",
+    "service-runner": "^2.7.6",
+    "speech-rule-engine": "^2.4.0",
+    "svgo": "^1.3.2",
+    "swagger-router": "^0.7.4",
+    "texvcinfo": "^0.5.5"
   },
   "optionalDependencies": {
     "librsvg": "^0.7.0",
-    "swagger-ui-dist": "^3.23.11"
+    "swagger-ui-dist": "^3.25.0"
   },
   "devDependencies": {
-    "ajv": "^6.6.2",
-    "commander": "^2.9.0",
-    "coveralls": "^3.0.2",
-    "dom-compare": "^0.3.1",
-    "eslint-config-wikimedia": "^0.11.0",
-    "eslint-plugin-jsdoc": "^4.1.1",
-    "eslint-plugin-json": "^1.4.0",
-    "eslint": "^5.15.1",
-    "extend": "^3.0.0",
+    "ajv": "^6.12.0",
+    "commander": "^5.0.0",
+    "coveralls": "^3.0.9",
+    "dom-compare": "^0.6.0",
+    "eslint-config-wikimedia": "^0.15.0",
+    "eslint-plugin-jsdoc": "^22.0.1",
+    "eslint-plugin-json": "^2.1.0",
+    "eslint": "^6.8.0",
+    "extend": "^3.0.2",
     "istanbul": "^0.4.5",
     "microtime": "^3.0.0",
-    "mocha": "^5.2.0",
+    "mocha": "^7.1.0",
     "mocha-lcov-reporter": "^1.3.0",
     "mock-require": "^3.0.3",
-    "nyc": "^13.3.0",
-    "rewire": "^4.0.1",
-    "xmldom": "^0.1.21"
+    "nyc": "^15.0.0",
+    "rewire": "^5.0.0",
+    "xmldom": "^0.3.0"
   },
   "engines": {
     "node": ">=10"
