@@ -175,7 +175,7 @@ function getPNG(result, resolve, conf) {
                 width: width * pngScale,
                 height: height * pngScale
             });
-            result.png = buffer.data || new Buffer();
+            result.png = buffer.data || Buffer.from([]);
         } catch (e) {
             result.errors = e.message;
         }
